+++ Specs/Scene/CameraSpec.js
@@ -945,6 +945,79 @@ defineSuite([
         }).toThrowDeveloperError();
     });
 
+    it('lookAtHeadingPitchRange', function() {
+        var target = Cartesian3.fromDegrees(0.0, 0.0);
+        var heading = CesiumMath.toRadians(45.0);
+        var pitch = CesiumMath.toRadians(-45.0);
+        var range = 2.0;
+
+        var tempCamera = camera.clone();
+        tempCamera.lookAtHeadingPitchRange(target, heading, pitch, range);
+
+        tempCamera.lookAtTransform(Matrix4.IDENTITY);
+
+        expect(Cartesian3.distance(tempCamera.position, target)).toEqualEpsilon(range, CesiumMath.EPSILON6);
+        expect(tempCamera.heading).toEqualEpsilon(heading, CesiumMath.EPSILON6);
+        expect(tempCamera.pitch).toEqualEpsilon(pitch, CesiumMath.EPSILON6);
+
+        expect(1.0 - Cartesian3.magnitude(tempCamera.direction)).toBeLessThan(CesiumMath.EPSILON14);
+        expect(1.0 - Cartesian3.magnitude(tempCamera.up)).toBeLessThan(CesiumMath.EPSILON14);
+        expect(1.0 - Cartesian3.magnitude(tempCamera.right)).toBeLessThan(CesiumMath.EPSILON14);
+    });
+
+    it('lookAtHeadingPitchRange throws with no target parameter', function() {
+        expect(function() {
+            camera.lookAtHeadingPitchRange(undefined, 1.0, 2.0, 3.0);
+        }).toThrowDeveloperError();
+    });
+
+    it('lookAtHeadingPitchRange throws with no heading parameter', function() {
+        expect(function() {
+            camera.lookAtHeadingPitchRange(Cartesian3.ZERO, undefined, 2.0, 3.0);
+        }).toThrowDeveloperError();
+    });
+
+    it('lookAtHeadingPitchRange throws with no pitch parameter', function() {
+        expect(function() {
+            camera.lookAtHeadingPitchRange(Cartesian3.ZERO, 1.0, undefined, 3.0);
+        }).toThrowDeveloperError();
+    });
+
+    it('lookAtHeadingPitchRange throws with no range parameter', function() {
+        expect(function() {
+            camera.lookAtHeadingPitchRange(Cartesian3.ZERO, 1.0, 2.0, undefined);
+        }).toThrowDeveloperError();
+    });
+
+    it('lookAtHeadingPitchRange in 2D mode', function() {
+        var frustum = new OrthographicFrustum();
+        frustum.near = 1.0;
+        frustum.far = 2.0;
+        frustum.left = -2.0;
+        frustum.right = 2.0;
+        frustum.top = 1.0;
+        frustum.bottom = -1.0;
+
+        var tempCamera = camera.clone();
+        tempCamera.frustum = frustum;
+        tempCamera.update(SceneMode.SCENE2D);
+
+        var target = Cartesian3.fromDegrees(0.0, 0.0);
+        var heading = CesiumMath.toRadians(90.0);
+        var pitch = CesiumMath.toRadians(-45.0);
+        var range = 2.0;
+
+        tempCamera.lookAtHeadingPitchRange(target, heading, pitch, range);
+
+        expect(Cartesian2.clone(tempCamera.position)).toEqual(Cartesian2.ZERO);
+
+        tempCamera.lookAtTransform(Matrix4.IDENTITY);
+        expect(tempCamera.direction).toEqual(Cartesian3.negate(Cartesian3.UNIT_Z, new Cartesian3()));
+        expect(tempCamera.heading).toEqualEpsilon(heading, CesiumMath.EPSILON6);
+        expect(tempCamera.frustum.right).toEqual(range);
+        expect(tempCamera.frustum.left).toEqual(-range);
+    });
+
     it('lookAtTransform', function() {
         var target = new Cartesian3(-1.0, -1.0, 0.0);
         var offset = new Cartesian3(1.0, 1.0, 0.0);
@@ -1021,6 +1094,89 @@ defineSuite([
 
         expect(function() {
             camera.lookAtTransform(Matrix4.IDENTITY, Cartesian3.UNIT_X);
+        }).toThrowDeveloperError();
+    });
+
+    it('lookAtTransformHeadingPitchRange', function() {
+        var target = Cartesian3.fromDegrees(0.0, 0.0);
+        var heading = CesiumMath.toRadians(45.0);
+        var pitch = CesiumMath.toRadians(-45.0);
+        var range = 2.0;
+        var transform = Transforms.eastNorthUpToFixedFrame(target);
+
+        var tempCamera = camera.clone();
+        tempCamera.lookAtTransformHeadingPitchRange(transform, heading, pitch, range);
+
+        tempCamera.lookAtTransform(Matrix4.IDENTITY);
+
+        expect(Cartesian3.distance(tempCamera.position, target)).toEqualEpsilon(range, CesiumMath.EPSILON6);
+        expect(tempCamera.heading).toEqualEpsilon(heading, CesiumMath.EPSILON6);
+        expect(tempCamera.pitch).toEqualEpsilon(pitch, CesiumMath.EPSILON6);
+
+        expect(1.0 - Cartesian3.magnitude(tempCamera.direction)).toBeLessThan(CesiumMath.EPSILON14);
+        expect(1.0 - Cartesian3.magnitude(tempCamera.up)).toBeLessThan(CesiumMath.EPSILON14);
+        expect(1.0 - Cartesian3.magnitude(tempCamera.right)).toBeLessThan(CesiumMath.EPSILON14);
+    });
+
+    it('lookAtTransformHeadingPitchRange throws with no transform parameter', function() {
+        expect(function() {
+            camera.lookAtTransformHeadingPitchRange(undefined, 1.0, 2.0, 3.0);
+        }).toThrowDeveloperError();
+    });
+
+    it('lookAtTransformHeadingPitchRange throws with no heading parameter', function() {
+        expect(function() {
+            camera.lookAtTransformHeadingPitchRange(Matrix4.IDENTITY, undefined, 2.0, 3.0);
+        }).toThrowDeveloperError();
+    });
+
+    it('lookAtTransformHeadingPitchRange throws with no pitch parameter', function() {
+        expect(function() {
+            camera.lookAtTransformHeadingPitchRange(Matrix4.IDENTITY, 1.0, undefined, 3.0);
+        }).toThrowDeveloperError();
+    });
+
+    it('lookAtTransformHeadingPitchRange throws with no range parameter', function() {
+        expect(function() {
+            camera.lookAtTransformHeadingPitchRange(Matrix4.IDENTITY, 1.0, 2.0, undefined);
+        }).toThrowDeveloperError();
+    });
+
+    it('lookAtTransformHeadinPitchRange in 2D mode', function() {
+        var frustum = new OrthographicFrustum();
+        frustum.near = 1.0;
+        frustum.far = 2.0;
+        frustum.left = -2.0;
+        frustum.right = 2.0;
+        frustum.top = 1.0;
+        frustum.bottom = -1.0;
+
+        var tempCamera = camera.clone();
+        tempCamera.frustum = frustum;
+        tempCamera.update(SceneMode.SCENE2D);
+
+        var target = Cartesian3.fromDegrees(0.0, 0.0);
+        var heading = CesiumMath.toRadians(90.0);
+        var pitch = CesiumMath.toRadians(-45.0);
+        var range = 2.0;
+        var transform = Transforms.eastNorthUpToFixedFrame(target);
+
+        tempCamera.lookAtTransformHeadingPitchRange(transform, heading, pitch, range);
+
+        expect(Cartesian2.clone(tempCamera.position)).toEqual(Cartesian2.ZERO);
+
+        tempCamera.lookAtTransform(Matrix4.IDENTITY);
+        expect(tempCamera.direction).toEqual(Cartesian3.negate(Cartesian3.UNIT_Z, new Cartesian3()));
+        expect(tempCamera.heading).toEqualEpsilon(heading, CesiumMath.EPSILON6);
+        expect(tempCamera.frustum.right).toEqual(range);
+        expect(tempCamera.frustum.left).toEqual(-range);
+    });
+
+    it('lookAtTransformHeadinPitchRange throws when morphing', function() {
+        camera.update(SceneMode.MORPHING);
+
+        expect(function() {
+            camera.lookAtTransformHeadingPitchRange(Matrix4.IDENTITY, 1.0, 2.0, 3.0);
         }).toThrowDeveloperError();
     });